feat: add `defineConfig` helper
