+++ lib/config.js
@@ -1,0 +1,1 @@
+export const defineConfig = (configuration) => configuration
